Bugs and improvements
