+++ contribute/src/components/LeftNav.js
@@ -55,20 +55,8 @@ const LeftNav = (props) => {
     const css = useStyles();
     const { path, url } = useRouteMatch();
 
-    const token = useSelector(state => state.auth.token);
     const stats = useSelector(state => state.requests.stats);
-    const dispatch = useDispatch();
-
-    useEffect(() => {
-        const Request = new DjangoRequest(token);
-        dispatch({ type: 'ASYNC.START '});
-        Request.stats()
-            .then((result) => {
-                dispatch({ type:'REQUESTS.STATS', payload: result });
-                dispatch({ type: 'ASYNC.STOP '});
-            });
-    }, []);
-
+    
     return (
         <Drawer
             variant="permanent"
@@ -93,7 +81,7 @@ const LeftNav = (props) => {
                             button 
                             to={`${url}requests/pending`}
                             primary="Pending"
-                            badge={stats.new && <Chip size="small" color="secondary" label={stats.new} />}
+                            badge={stats.new > 0 && <Chip size="small" color="secondary" label={stats.new} />}
                         />
                         {/*
                         <ListItemLink
@@ -107,7 +95,7 @@ const LeftNav = (props) => {
                             button 
                             to={`${url}requests/draft`}
                             primary="Drafts"
-                            badge={stats.drafts && <Chip size="small" color="secondary" label={stats.drafts} />}
+                            badge={stats.drafts > 0 && <Chip size="small" color="secondary" label={stats.drafts} />}
                         />
                     </List>
                 </List>
