+++ contribute/src/components/MainContent.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useEffect} from 'react';
 
 import LeftNav from './LeftNav'
 import ListRequests from './ListRequests'
@@ -8,6 +8,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import Snackbar from "@material-ui/core/Snackbar";
 import Alert from "@material-ui/lab/Alert";
 import {useDispatch, useSelector} from "react-redux";
+import DjangoRequest from '../objects/DjangoRequest';
 
 
 const useStyles = makeStyles((theme) => ({
@@ -23,6 +24,9 @@ const MainContent = (props) => {
     const css = useStyles();
     const { path, url } = useRouteMatch();
 
+    // Auth token
+    const token = useSelector(state => state.auth.token);
+
     // Redux
     const dispatch = useDispatch();
 
@@ -33,6 +37,21 @@ const MainContent = (props) => {
     const handleSbClose = () => {
         dispatch({ type:'ASYNC.SB_CLOSE' });
     }
+
+    // Stats
+    const loadStats = () => {
+        const Request = new DjangoRequest(token);
+        Request.stats()
+            .then((result) => {
+                dispatch({ type:'REQUESTS.STATS', payload: result });
+            });
+    }
+    useEffect(() => {
+        loadStats();
+        const interval = setInterval(loadStats, 30000);
+        return () => clearInterval(interval);
+    }, []);
+
 
     return (
         <>
